requirements.txt
apache-libcloud==2.4.0
Flask==1.0.2
Flask-Cors==3.0.6
Flask-HTTPAuth==3.2.4
gevent==1.3.5
gunicorn==19.9.0
mongoengine==0.15.0
pymongo==3.6.1
